+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/util/CaseListUtils.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util
 
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.ManagedCase
+import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.ManagedCase
 import java.time.Clock
 import java.time.LocalDate
